+++ include/ignition/gazebo/components/JointVelocityCmd.hh
@@ -34,8 +34,9 @@ inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
   /// \brief Base class which can be extended to add serialization
-  using JointVelocityCmd = Component<std::vector<double>,
-        class JointVelocityCmdTag>;
+  using JointVelocityCmd =
+      Component<std::vector<double>, class JointVelocityCmdTag,
+                serializers::VectorDoubleSerializer>;
 
   IGN_GAZEBO_REGISTER_COMPONENT(
       "ign_gazebo_components.JointVelocityCmd", JointVelocityCmd)